move SplitPaneState and SplitterState classes to splitpane package
change internal backed fields of SplitterState from var to val
